+++ src/graphql/provider/graphql-ratchet-endpoint-provider.ts
@@ -1,0 +1,3 @@
+export interface GraphqlRatchetEndpointProvider {
+  fetchGraphqlEndpoint(): string;
+}