+++ src/graphql/provider/graphql-ratchet-error-handler.ts
@@ -1,0 +1,3 @@
+export interface GraphqlRatchetErrorHandler {
+  handleError(error: any, queryName: string, variables: Record<string, any>, anonymous: boolean): void;
+}