+++ src/graphql/provider/graphql-ratchet-query-provider.ts
@@ -1,0 +1,3 @@
+export interface GraphqlRatchetQueryProvider {
+  fetchQueryText(name: string): Promise<string>;
+}